Back_end/requirements.txt
flask
flask-cors
python-dotenv
google-genai
Pillow
gunicorn